requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
boto3==1.35.90
openpyxl==3.1.5
pillow==11.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
